RayTracingWeekend

Documentation for RayTracingWeekend.

RayTracingWeekend.Ray

Base.:*

Base.:+

Base.:-

Base.:-

Base.:/

Base.:==

Base.print

RayTracingWeekend.:×

RayTracingWeekend.:⋅

RayTracingWeekend.at

RayTracingWeekend.len

RayTracingWeekend.len_squared

RayTracingWeekend.pixel_color

RayTracingWeekend.random_vector

RayTracingWeekend.sky_color

RayTracingWeekend.to_point

RayTracingWeekend.to_vec

RayTracingWeekend.Ray — Type

r = Ray(origin, direction)
r.origin ::Point 
r.direction ::Vec 
at(r::Ray, t::Float32) ::Point

Rays simulate the physical rays, they have a direction and origin. The value t will give you a point along the ray in positive and negative direction.

Base.:* — Method

Real * p::XYZ -> p::XYZ

Return a XYZ (Point, Vec) multiplied by a real.

x = Vec{Int32}(1.0, 2.0, 3.0) x2 = x * 2 x2 = 2 * x

Base.:+ — Method

p3::AbstractXYZ = p1::AbstractXYZ + p2::AbstractXYZ

Adds p1 and p2 and returns a new AbstractXYZ (Point, Vec).

There are 3 cases to consider:

Addition of 2 points

Addition of a point and vec

Addition of 2 vectors

Addition is not defined for XYZ type.

x1 = Point{Float64}(1.0, 2.0, 3.0) x2 = Vec{Int32}(1.0, 2.0, 3.0) print(x1 + x2)

x1 = Vec{Float64}(1.0, 2.0, 3.0) x2 = Vec{Int32}(1.0, 2.0, 3.0) print(x1 + x2)

x1 = Point{Int32}(1.0, 2.0, 3.0) x2 = Point{Int32}(1.0, 2.0, 3.0) print(x1 + x2)

Base.:- — Method

p3::AbstractXYZ = p1::AbstractXYZ - p2:AbstractXYZ

Subtract p1 from p2, and return a new (Point, Vec).

Check addition docs to determine for more details.

Base.:- — Method

- p::XYZ -> -p::XYZ

Return a negative value of the given XYZ (Point, Vec).

Base.:/ — Method

p::AbstractXYZ / r::Real -> p::AbstractXYZ

Returns a AbstractXYZ (Point, Vec) divided by a real r.

x = Vec{Int32}(1.0, 2.0, 3.0) div_x = x/2.0 Vec{Float64}(0.5, 1.0, 1.5)

Base.:== — Method

p1::AbstractXYZ == p2::AbstractXYZ

Check whether the given 2 AbstractXYZ (Point, Vec) are equal.

Base.print — Method

print(p<:AbstractXYZ) Prints the struct (Point, Vec) with values.

x = XYZ{Float64}(1.0, 2.0, 3.0) print(x)

x = Point{Float64}(1.0, 2.0, 3.0) print(x)

x = Vec{Float64}(1.0, 2.0, 3.0) print(x)

RayTracingWeekend.:× — Method

p1::AbstractVec × p2::AbstractVec -> ::AbstractVec

Cross product of 2 vectors, return a new vector

RayTracingWeekend.:⋅ — Method

p1::AbstractVec ⋅ p2::AbstractVec -> Float32

Returns the result of dot product of 2 Vec

x1 = Vec{Int32}(1.0, 2.0, 3.0) x2 = Vec{Int32}(1.0, 2.0, 3.0) println(x1 ⋅ x2)

RayTracingWeekend.at — Method

at(ray, t) returns the point in 3D of ray at t.

RayTracingWeekend.len — Method

len(p1::AbstractXYZ) -> Real

Returns the √(x² + y² + z²)

RayTracingWeekend.len_squared — Method

len_squared(p1::XYZ) -> Real

Returns x² + y² + z²

RayTracingWeekend.pixel_color — Method

pixel_color (p::Point) -> ::Point{Int8} Return an RGB pixel color for the input point p where p.x, p.y and p.z is 0-1 range.

RayTracingWeekend.random_vector — Method

random vector with values between min and max

RayTracingWeekend.sky_color — Method

The skycolor(ray) function linearly blends white and blue depending on the height of the y coordinate after scaling the ray direction to unit length. skycolor(r::Ray) -> ::RGB{Float32}

RayTracingWeekend.to_point — Method

Convert a vec to a point

RayTracingWeekend.to_vec — Method

Convert a point to a vector (subtract from Vec(0.0, 0.0, 0.0))