RayTracingWeekend

Documentation for RayTracingWeekend.

RayTracingWeekend.Camera

RayTracingWeekend.HitRecord

RayTracingWeekend.HitRecord2

RayTracingWeekend.Ray

RayTracingWeekend.Sphere

RayTracingWeekend.Sphere2

RayTracingWeekend.:×

RayTracingWeekend.:⋅

RayTracingWeekend.at

RayTracingWeekend.get_camera

RayTracingWeekend.get_hit_record

RayTracingWeekend.get_hit_record

RayTracingWeekend.get_ray

RayTracingWeekend.get_sampled_color

RayTracingWeekend.hit

RayTracingWeekend.hit!

RayTracingWeekend.len

RayTracingWeekend.len_squared

RayTracingWeekend.pixel_color

RayTracingWeekend.random_vector

RayTracingWeekend.set_hit_record

RayTracingWeekend.set_hit_record

RayTracingWeekend.set_hit_record

RayTracingWeekend.set_hit_record

RayTracingWeekend.sky_color

RayTracingWeekend.to_point

RayTracingWeekend.to_vec

RayTracingWeekend.Camera — Type

Camera object with specification for origin, horizontal and vertical vectors. We can use the camera object to generate rays along each (u,v) pixels.

RayTracingWeekend.HitRecord — Type

HitRecord stores information on the point of impact for a ray.
This includes where ray hits, "t" in o+d(t), surface normal 
and where the impact was like the front side or back side.

RayTracingWeekend.HitRecord2 — Type

HitRecord2 stores material inaddition to details present in HitRecord.

RayTracingWeekend.Ray — Type

r = Ray(origin, direction)
r.origin ::Point 
r.direction ::Vec 
at(r::Ray, t::Float32) ::Point

Rays simulate the physical rays, they have a direction and origin. The value t will give you a point along the ray in positive and negative direction.

RayTracingWeekend.Sphere — Type

Sphere stores information on center and radius.

RayTracingWeekend.Sphere2 — Type

Sphere2 stores material inaddition to center and radius.

RayTracingWeekend.:× — Method

p1::AbstractVec × p2::AbstractVec -> ::AbstractVec

Cross product of 2 vectors, return a new vector

RayTracingWeekend.:⋅ — Method

p1::AbstractVec ⋅ p2::AbstractVec -> Float32

Returns the result of dot product of 2 Vec

x1 = Vec{Int32}(1.0, 2.0, 3.0) x2 = Vec{Int32}(1.0, 2.0, 3.0) println(x1 ⋅ x2)

RayTracingWeekend.at — Method

at(ray, t) returns the point in 3D of ray at t.

RayTracingWeekend.get_camera — Method

Create a camera object from aspectratio, viewportheight and focal_length

RayTracingWeekend.get_hit_record — Method

Get a HitRecord2 object with default values, and a given material (since material is abstract).

RayTracingWeekend.get_hit_record — Method

Get a HitRecord object with default values.

RayTracingWeekend.get_ray — Method

Create a ray based on u, v inputs on the camera

RayTracingWeekend.get_sampled_color — Method

get_sampled_color applies gamma correction and scales the pixel color by the number of samples taken.

RayTracingWeekend.hit! — Method

hit! function for sphere, calculates the impact location, normal etc on HitRecord `rec`
for the `ray`. 
- This function will update values on rec on a successful hit.

RayTracingWeekend.hit — Method

Apply ray hit on all the objects in the scene.

RayTracingWeekend.len — Method

len(p1::AbstractXYZ) -> Real

Returns the √(x² + y² + z²)

RayTracingWeekend.len_squared — Method

len_squared(p1::XYZ) -> Real

Returns x² + y² + z²

RayTracingWeekend.pixel_color — Method

pixel_color (p::Point) -> ::Point{Int8} Return an RGB pixel color for the input point p where p.x, p.y and p.z is 0-1 range.

RayTracingWeekend.random_vector — Method

random vector with values between min and max

RayTracingWeekend.set_hit_record — Method

Set the source HitRecord to be same as the destination HitRecord.
Equivalent of copy constructor in C++. (Harish: maybe slow?)

RayTracingWeekend.set_hit_record — Method

Set HitRecord `rec` to the arguments passed.

RayTracingWeekend.set_hit_record — Method

Set the source HitRecord2 to be same as the destination HitRecord2.
Equivalent of copy constructor in C++. (Harish: maybe slow?)

RayTracingWeekend.set_hit_record — Method

sethitrecord sets fields inside HitRecord2 rec to the arguments passed

RayTracingWeekend.sky_color — Method

The skycolor(ray) function linearly blends white and blue depending on the height of the y coordinate after scaling the ray direction to unit length. skycolor(r::Ray) -> ::RGB{Float32}

RayTracingWeekend.to_point — Method

Convert a vec to a point

RayTracingWeekend.to_vec — Method

Convert a point to a vector (subtract from Vec(0.0, 0.0, 0.0))